requirements.txt
PyPDF2
Flask
pytesseract
pillow
python-dateutil
transformers[torch]
python-doctr
requests
python-dotenv